preparations to make a single command buffer

+++ crates/ris_video/src/vulkan/mod.rs
@@ -32,7 +32,7 @@ const PREFERRED_FORMAT: vk::Format = vk::Format::B8G8R8A8_SRGB;
 const PREFERRED_COLOR_SPACE: vk::ColorSpaceKHR = vk::ColorSpaceKHR::SRGB_NONLINEAR;
 const PREFERRED_PRESENT_MODE: vk::PresentModeKHR = vk::PresentModeKHR::IMMEDIATE;
 
-const MAX_FRAMES_IN_FLIGHT: u32 = 2;
+const MAX_FRAMES_IN_FLIGHT: usize = 2;
 
 pub const VERTICES: [Vertex; 4 * 6] = [
     // pos x
